Add the windows-x86_64-msvc target
OpenSSL VC-WIN64A no-shared and FreeRDP with Ninja on the dynamic CRT, the import-library
set measured against a cl-linked probe, a natively generated bindings_windows.rs with a
kernel32 shim for what the SDK-guarded WinPR headers drop, and the ci/windows driver for
the windows-ci-build box.

+++ crates/freerdp/src/session.rs
@@ -1906,7 +1906,10 @@ unsafe extern "C" fn surface_frame_marker(
         let Some(bridge) = (unsafe { bridge(ctx) }) else { return 0 };
         // SAFETY: as `frame_marker`; `update` and `settings` are live on a connected context.
         unsafe {
-            if (*marker).frameAction != sys::SURFCMD_FRAMEACTION_SURFACECMD_FRAMEACTION_END {
+            // `as u32`: the enum constant is `c_uint` from clang and `c_int` from MSVC; the
+            // field is a UINT32 on both.
+            #[allow(clippy::unnecessary_cast)]
+            if (*marker).frameAction != sys::SURFCMD_FRAMEACTION_SURFACECMD_FRAMEACTION_END as u32 {
                 return 1;
             }
             if sys::freerdp_settings_get_uint32(
